requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.2
psycopg2-binary==2.9.5
python-dotenv==1.0.0
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
SQLAlchemy==2.0.7
threadpoolctl==3.1.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.2
